# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApproverOut": {
        "additionalProperties": false,
        "description": "Deliberately narrower than `UserOut` \u2014 a cashier looking up who can\napprove an override needs a name to pick, not a colleague's phone\nnumber, email, or role listing.",
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name"
        ],
        "title": "ApproverOut",
        "type": "object"
      },
...
      "GrantRequest": {
        "additionalProperties": false,
        "properties": {
          "effect": {
            "title": "Effect",
            "type": "string"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "permission_key": {
            "title": "Permission Key",
            "type": "string"
          }
        },
        "required": [
          "permission_key",
          "effect"
        ],
        "title": "GrantRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MeOut": {
        "additionalProperties": false,
        "properties": {
          "business_id": {
            "title": "Business Id",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "location_ids": {
            "items": {
              "type": "string"
            },
            "title": "Location Ids",
            "type": "array"
          },
          "role_key": {
            "title": "Role Key",
            "type": "string"
          }
        },
        "required": [
          "id",
          "business_id",
          "display_name",
          "role_key",
          "location_ids"
        ],
        "title": "MeOut",
        "type": "object"
      },
...
      "RoleChangeRequest": {
        "additionalProperties": false,
        "properties": {
          "role_key": {
            "title": "Role Key",
            "type": "string"
          }
        },
        "required": [
          "role_key"
        ],
        "title": "RoleChangeRequest",
        "type": "object"
      },
...
      "UserCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "location_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Location Ids",
            "type": "array"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "role_key": {
            "title": "Role Key",
            "type": "string"
          },
          "secret": {
            "title": "Secret",
            "type": "string"
          }
        },
        "required": [
          "display_name",
          "secret",
          "role_key"
        ],
        "title": "UserCreateRequest",
        "type": "object"
      },
      "UserOut": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "role_key": {
            "title": "Role Key",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name",
          "phone",
          "email",
          "role_key",
          "status"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "OperatorOS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "operationId": "list_users_api_v1_users_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserOut"
                  },
                  "title": "Response List Users Api V1 Users Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Users",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "create_user_api_v1_users_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create User",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/approvers": {
      "get": {
        "description": "Who can approve an override that the current user isn't allowed to\nmake alone -- a discount above the threshold, a credit-limit override, a\nback-dated payment.\n\n`sales.py::_verify_manager_override` needs BOTH a manager's user id and\nthat manager's PIN, so the Counter has to be able to name the approver.\n`GET /users` can't serve that: it requires `user.manage`, which a\ncashier deliberately does not have, so the frontend previously sent\n`manager_override_user_id: null` and every over-threshold sale failed\n422 no matter what PIN was typed.\n\nDeliberately NOT capability-gated beyond being authenticated: a cashier\nmust be able to see who to ask. It returns only id and display name --\nno phone, no email, no role listing -- for active users in the caller's\nown business who hold the specific capability being requested. That is\nstrictly less than the shop's own staff already know about each other.",
        "operationId": "list_approvers_api_v1_users_approvers_get",
        "parameters": [
          {
            "in": "query",
            "name": "capability",
            "required": true,
            "schema": {
              "title": "Capability",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApproverOut"
                  },
                  "title": "Response List Approvers Api V1 Users Approvers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Approvers",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me": {
      "get": {
        "operationId": "get_me_api_v1_users_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Me",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{user_id}": {
      "get": {
        "operationId": "get_user_api_v1_users__user_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get User",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{user_id}/grants": {
      "post": {
        "description": "Writes a PERMISSION_OVERRIDDEN audit_log entry -- the per-user\ngrant/revoke layer described in capabilities.py, exposed as a real\n(if minimal) endpoint so it's exercised end-to-end rather than only\nat the function level.",
        "operationId": "override_user_permission_api_v1_users__user_id__grants_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Override User Permission",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{user_id}/role": {
      "post": {
        "description": "Writes a ROLE_CHANGED audit_log entry (spec G.1 / approved plan \u00a76).\nThis is deliberately minimal -- no feature UI calls it yet -- but it's\na real, capability-gated, RLS-protected mutation, not a stub: it's how\nthe audit log actually receives a ROLE_CHANGED event rather than one\nonly ever asserted about in a unit test.",
        "operationId": "change_user_role_api_v1_users__user_id__role_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change User Role",
        "tags": [
          "users"
        ]
      }
    },
...
  }
}